requirements.txt
azure-eventhub>=5.15.1
azure-identity>=1.12.0
Flask>=2.2.5
gunicorn[gevent]
gevent
